pass mealsArray as a props to MealsSearch, search updates after adding newMeal
