+++ src/components/MealsSearch/MealsSearch.module.css
@@ -1,36 +1,4 @@
-.groupStyles {
-  display: "flex";
-  align-items: "center";
-  justify-content: "space-between";
-}
-.groupBadgeStyles {
-  background-color: "#EBECF0";
-  border-radius: "2em";
-  color: "#172B4D";
-  display: "inline-block";
-  font-size: 12;
-  font-weight: "normal";
-  line-height: "1";
-  min-width: 1;
-  padding: "0.16666666666667em 0.5em";
-  text-align: "center";
-}
-
 .selectWrapper {
   width: 80%;
   max-height: 10%;
 }
-/* import React from "react";
-
-import Select from "react-select";
-import { colourOptions, groupedOptions } from "./data";
-import styles from "./MealsSearch.module.css";
-
-const { groupBadgeStyles, groupStyles } = styles;
-
-class MealsSearch extends React.Component {
-  render() {
-    return <div />;
-  }
-}
-export default MealsSearch; */
